feat: implementar app.js en todas las paginas
Se agrega el app.js en las paginas, junto a los ids y respectivos eventos onClick en el boton de Forma Parte.
Ademas, se movio el app.js a la carpeta de /js
