+++ página web/contacto.html
@@ -4,6 +4,7 @@
 <head>
     <title>Contacto</title>
     <script src="https://code.jquery.com/jquery-3.6.0.js" integrity="sha256-H+K7U5CnXl1h5ywQfKtSj8PCmoN9aaq30gDh27Xc0jk=" crossorigin="anonymous"></script>
+    <script type="module" src="js/app.js"></script>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Jost&display=swap" rel="stylesheet"> 
@@ -34,8 +35,8 @@
             <a class="item" href="guia.html">Guías</a>
             <a class="item" href="galeria.html">Galería</a>
             <a class="item" href="contacto.html">Contacto</a>
-            <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal">Forma parte</button>
-        </div>
+            <button id="botonAuth" onclick="click()" type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal">Forma parte</button>
+          </div>
       </nav>
       <nav class="navbar bg-light fixed-top navipad">
         <div class="container-fluid">
