+++ página web/index.html
@@ -5,6 +5,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <script type="module" src="js/app.js"></script>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Jost&display=swap" rel="stylesheet"> 
@@ -37,7 +38,7 @@
                 <a class="item" href="guia.html">Guías</a>
                 <a class="item" href="galeria.html">Galería</a>
                 <a class="item" href="contacto.html">Contacto</a>
-                <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal">Forma parte</button>
+                <button id="botonAuth" onclick="click()" type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal">Forma parte</button>
             </div>
           </nav>
           <nav class="navbar bg-light fixed-top navipad">
